feat(HomeBrew): AND-1353 BigDecimal moshi adaptor and tests plus use in a DTO (#1155)
* feat(HomeBrew): AND-1353 BigDecimal moshi adaptor and tests
* feat(HomeBrew): AND-1353 BigDecimal in use in trading config DTO
* feat(HomeBrew): AND-1353 Tidier moshiInterceptor Koin extension

+++ nabu/src/main/kotlin/com/blockchain/nabu/service/NabuMarketsService.kt
@@ -17,7 +17,7 @@ class NabuMarketsService internal constructor(
                 tradingPair.pairCodeUpper,
                 it.authHeader
             ).map {
-                TradingConfig(minOrderSize = CryptoValue.fromMajor(tradingPair.from, it.minOrderSize.toBigDecimal()))
+                TradingConfig(minOrderSize = CryptoValue.fromMajor(tradingPair.from, it.minOrderSize))
             }
         }
     }
